+++ www/whatsnew.md
@@ -12,6 +12,7 @@ Bug fixes.
 
  - Klive Icetooth now correctly plays its Opener. (1124)
  - The End now applies to both player's hands and the stats effect applies before the Opener is evaluated. (1126)
+ - Uccian Hydra now correctly triggers. (1125)
 
 ### 0.8.31-2.0.30 (Tuesday, May 28th, 2019)